feat(brokerage): Phase 3 — XLSX import, prices, benchmarks, missing accts
This commit lands T4-T19 of the brokerage Phase 3 pipeline (`/qpipeline 005-page`),
building on the Phase 3 foundation in 0a73c3a. Total: 1550 tests passing
(76 net new across the phase).

Adapters & scripts (T4, T8, T11, T16, T17):
- src/adapters/xlsx_savings_plan.py — XLSX importer for Account Summary
  (~30 accounts × 36 dates = 500 candidate balance snapshots), Historical
  Prices seed (74 candidates × 18 symbols), TD GainLoss + SB Raw lots
  (368 + 96 = 464 lot rows). DRY-RUN by default; CLI subcommands
  import-balances, import-prices, import-lots.
- src/adapters/yfinance_prices.py — yfinance EOD adapter behind a
  HistoricalPriceRow contract; mocked in tests, no live API calls.
- scripts/backfill_historical_prices.py — bulk backfill from yfinance,
  idempotent on (symbol, trade_date) PK; discovers symbols from
  position_snapshot + brokerage_transaction; defaults SPY/VTI/QQQ/BND
  benchmarks.
- scripts/seed_expected_accounts.py — seeds expected_account from XLSX
  + Credit Karma list (17 + 18 rows), auto-links by last_4 (12 matches),
  interactive `confirm` walks unconfirmed rows with active/closed prompts.

API endpoints (T6, T12, T14, T18):
- /api/brokerage/networth-history — aggregated balance series across
  matched accounts; excludes closed expected_account rows.
- /api/brokerage/networth-history-benchmark?benchmark=SPY — buy-and-hold
  simulation aligned with portfolio dates.
- /api/brokerage/holdings/{symbol}/history — per-symbol value series
  + lot-level cost basis from cost_basis_lot.
- /api/brokerage/missing-accounts — active expected_accounts whose
  resolved Account has no snapshot newer than 60 days (configurable).

Dashboard (T7, T13, T15, T19):
- Net-worth-over-time SVG line chart on /brokerage with hover dots
  and date+value tooltip.
- "Compare to S&P 500" toggle: dashed orange overlay on the same chart
  with normalized buy-and-hold simulation; deltaPct shown in headline.
- Missing Accounts panel surfaces active accounts with stale or no
  snapshots; institution / account / last_4 / last_seen.
- /brokerage/holdings/[symbol] — per-holding detail page with 3M/6M/YTD/
  1Y/All range toggle, value chart, headline (qty / cost / unrealized),
  and Historical Lots table from XLSX import.

Schema fix:
- src/models/history.py — corrected `Mapped[float]` to `Mapped[Decimal]`
  on Numeric(asdecimal=True) columns for mypy correctness.

Operations:
- com.sparkry.accounting-prices-daily.plist — launchd plist runs the
  backfill incrementally Mon-Fri at 17:30 PT.
- pyproject.toml — added yfinance>=0.2.40 dep.

Pipeline state: /qpipeline 005-page advances from execute → review-loop.

+++ src/models/history.py
@@ -14,6 +14,7 @@ Decimal precision matches src/models/brokerage.py:
 
 import uuid
 from datetime import UTC, date, datetime
+from decimal import Decimal
 
 from sqlalchemy import (
     CheckConstraint,
@@ -57,16 +58,16 @@ class HistoricalPrice(Base):
 
     symbol: Mapped[str] = mapped_column(String(32), primary_key=True)
     trade_date: Mapped[date] = mapped_column(Date, primary_key=True)
-    close: Mapped[float] = mapped_column(
+    close: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=False
     )
-    open: Mapped[float | None] = mapped_column(
+    open: Mapped[Decimal | None] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=True
     )
-    high: Mapped[float | None] = mapped_column(
+    high: Mapped[Decimal | None] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=True
     )
-    low: Mapped[float | None] = mapped_column(
+    low: Mapped[Decimal | None] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=True
     )
     volume: Mapped[int | None] = mapped_column(nullable=True)
@@ -104,7 +105,7 @@ class AccountBalanceSnapshot(Base):
     )
     raw_account_name: Mapped[str] = mapped_column(String(255), nullable=False)
     as_of: Mapped[date] = mapped_column(Date, nullable=False)
-    balance: Mapped[float] = mapped_column(
+    balance: Mapped[Decimal] = mapped_column(
         Numeric(precision=14, scale=2, asdecimal=True), nullable=False
     )
     source: Mapped[str] = mapped_column(String(32), nullable=False)
@@ -175,16 +176,16 @@ class CostBasisLot(Base):
     symbol: Mapped[str] = mapped_column(String(32), nullable=False)
     security_name: Mapped[str | None] = mapped_column(String(255), nullable=True)
     open_date: Mapped[date] = mapped_column(Date, nullable=False)
-    quantity: Mapped[float] = mapped_column(
+    quantity: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=False
     )
-    cost_per_share: Mapped[float] = mapped_column(
+    cost_per_share: Mapped[Decimal] = mapped_column(
         Numeric(precision=18, scale=8, asdecimal=True), nullable=False
     )
-    cost_total: Mapped[float] = mapped_column(
+    cost_total: Mapped[Decimal] = mapped_column(
         Numeric(precision=14, scale=2, asdecimal=True), nullable=False
     )
-    wash_sale_adj: Mapped[float | None] = mapped_column(
+    wash_sale_adj: Mapped[Decimal | None] = mapped_column(
         Numeric(precision=14, scale=2, asdecimal=True), nullable=True
     )
     source: Mapped[str] = mapped_column(String(32), nullable=False)
